apps/api/pyproject.toml
[project]
name = "forgeai-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "sqlmodel>=0.0.22",
  "alembic>=1.13.3",
  "psycopg[binary]>=3.2.3",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.5.2",
  "python-multipart>=0.0.12"
]

[tool.uvicorn]
factory = false
